Add check for existing files.

+++ ftml/scripts/ast_stub.sh
@@ -8,6 +8,11 @@ set -eu
 cd "${0%/*}/.."
 
 for name in "$@"; do
+	if [[ -f test/$name.json ]] || [[ -f test/$name.txt ]] || [[ -f test/$name.html ]]; then
+		echo "Already exists: $name..."
+		continue
+	fi
+
 	echo "Creating $name..."
 
 	cp misc/ast-test-template.json "test/$name.json"
